refactor: member.service 패키지 내 코드 수정
