+++ 리브/server/src/main/java/umc/server/domain/member/service/query/MemberQueryService.java
@@ -1,4 +1,6 @@
 package umc.server.domain.member.service.query;
 
 public interface MemberQueryService {
+
+    void changeUsername(Long memberId, String newUsername);
 }
