+++ 리브/server/src/main/java/umc/server/domain/member/service/query/MemberQueryServiceImpl.java
@@ -1,4 +1,22 @@
 package umc.server.domain.member.service.query;
 
-public class MemberQueryServiceImpl {
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import umc.server.domain.member.repository.MemberRepository;
+
+@Service
+@RequiredArgsConstructor
+public class MemberQueryServiceImpl implements MemberQueryService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    @Transactional
+    public void changeUsername(Long memberId, String newUsername) {
+        int updated = memberRepository.updateUsername(memberId, newUsername);
+        if (updated == 0) {
+            throw new IllegalArgumentException("업데이트 대상 회원이 없습니다.");
+        }
+    }
 }
